pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "translate-chat"
version = "2.0.2"
# 最后更新(Updated): 2025/07/29
description = "实时语音识别与翻译桌面应用"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "深圳王哥 & AI", email = "support@translate-chat.com"}
]
keywords = ["语音识别", "翻译", "实时", "桌面应用", "kivy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "kivy>=2.3.0,<3.0.0",
    "kivymd==1.1.1",
    "pyaudio>=0.2.11,<0.3.0",
    "websocket-client>=1.6.0,<2.0.0",
    "aiohttp>=3.8.0,<4.0.0",
    "cryptography>=3.4.8,<4.0.0",
    "requests>=2.28.0,<3.0.0",
    "urllib3>=1.26.0,<2.0.0",
    "numpy>=1.21.0,<2.0.0",
    "scipy>=1.7.0,<2.0.0",
    "webrtcvad>=2.0.10,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
build = [
    "pyinstaller>=5.13.0",
    "setuptools>=61.0",
    "wheel>=0.37.0",
]

[project.urls]
Homepage = "https://github.com/manwjh/Translate-Chat"
Repository = "https://github.com/manwjh/Translate-Chat.git"
Documentation = "https://github.com/manwjh/Translate-Chat/wiki"
"Bug Tracker" = "https://github.com/manwjh/Translate-Chat/issues"

[project.scripts]
translate-chat = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "docs*", "scripts*", "releases*", "venv*", "build*", "dist*", ".dist*"]

[tool.setuptools.package-data]
"*" = ["*.py", "*.json", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short" 